move reused components to components folder
