+++ src/lib/components/DepartingTrainAnimation.svelte
