+++ src/lib/components/LiveTripAnimation.svelte
